Separating home layout for each locale.
