+++ locales/en/common.json
@@ -5,5 +5,6 @@
   "life": "Life",
   "estonia": "Estonia",
   "personal": "Personal",
+  "featured": "Featured article",
   "see-all": "See all"
 }
