+++ locales/id/common.json
@@ -5,5 +5,6 @@
   "life": "Kehidupan",
   "estonia": "Estonia",
   "personal": "Pribadi",
+  "featured": "Artikel pilihan",
   "see-all": "Lihat semua"
 }
