+++ public/rss.xml
@@ -4,7 +4,7 @@
         <title>The blog of Asep Bagja</title>
         <link>https://www.asepbagja.com</link>
         <description>The personal blog where Asep shares his opinion and topic that he is interested.</description>
-        <lastBuildDate>Wed, 02 Mar 2022 21:05:30 GMT</lastBuildDate>
+        <lastBuildDate>Thu, 03 Mar 2022 09:08:26 GMT</lastBuildDate>
         <docs>https://validator.w3.org/feed/docs/rss2.html</docs>
         <generator>The blog of Asep Bagja</generator>
         <language>en</language>
